+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change log for AutoPilotReadiness.ps1
 
+## 2.1.0 - August 1, 2023
+
+- Fixed license check for users; it checked it all the time. 
+- Fixed graph call for MDM; always checking .com
+- Fixed group check during user and mdm; remove security filter
+- Fixed graph scopes. Missing Policy.Read.All. 
+
 ## 2.0.0 - July 23, 2023
 
 - Added Intune license check; validates Intune license against service plans